User Story 1195736: UL: UDM integration - update api to send product id when disabling a product for a PMC

Cherry picked from !152476

Related work items: #1195736

+++ Enterprise/Subsystem/ProductLauncher/Service/LandingAPI/Controllers/OrganizationController.cs
@@ -1852,7 +1852,7 @@ namespace RP.Enterprise.Subsystem.ProductLauncher.Service.LandingAPI.Controllers
                     Id = org.PartyId,
                     CompanyInstanceSourceId = org.RealPageId.ToString().ToLower(),
                     CreatedBy = ProductEnumHelper.StringValueOf(ProductEnum.UnifiedPlatform) + " Automation",
-                    ProductCenterSourceId = product.ToString(),
+                    ProductCenterSourceId = ((int)product).ToString(),
                     PropertyInstanceSourceId = null,
                     Source = ProductEnumHelper.StringValueOf(ProductEnum.UnifiedPlatform)
